refactor(Cars): Car 리스트 일급컬렉션 적용
