+++ src/main/java/domain/RacingGame.java
@@ -11,12 +11,12 @@ public class RacingGame {
     private final static int GAME_OVER_COUNT = 0;
     private static final int MIN_GAME_TRIAL_COUNT = 1;
 
-    private final List<Car> cars;
+    private final Cars cars;
     private final NumberGenerator numberGenerator;
 
     private int gameTrialCount;
 
-    public RacingGame(List<Car> cars, int gameTrialCount, NumberGenerator numberGenerator) {
+    public RacingGame(Cars cars, int gameTrialCount, NumberGenerator numberGenerator) {
         validate(gameTrialCount);
         this.cars = cars;
         this.gameTrialCount = gameTrialCount;
@@ -29,17 +29,14 @@ public class RacingGame {
         }
     }
 
-    public List<Car> run() {
+    public Cars run() {
         start();
         finish();
-        return Collections.unmodifiableList(cars);
+        return cars;
     }
 
     public List<Car> getWinners() {
-        int maxMoveCount = getMaxMoveCount();
-        return cars.stream()
-                .filter(car -> car.isWinner(maxMoveCount))
-                .collect(Collectors.toList());
+        return cars.getWinners();
     }
 
     public boolean canContinue() {
@@ -47,19 +44,10 @@ public class RacingGame {
     }
 
     private void start() {
-        cars.stream()
-                .filter(car -> car.canMove(numberGenerator.generate()))
-                .forEach(Car::move);
+        cars.move(numberGenerator);
     }
 
     private void finish() {
         gameTrialCount--;
-    }
-
-    private int getMaxMoveCount() {
-        return cars.stream()
-                .mapToInt(Car::getMoveCount)
-                .max()
-                .getAsInt();
     }
 }
